feat: refactor typed_terms to terms and remove unused typed_terms module
- Updated imports across multiple files to replace `typed_terms` with `terms`.
- Refactored function calls and class instantiations to use the new `terms` module.
- Removed the `typed_terms.py` file as it was no longer needed.
- Adjusted related tests to ensure compatibility with the new structure.
